+++ src/parser/CronParser.ts
@@ -1,7 +1,7 @@
 import { CstParser, TokenVocabulary } from "chevrotain";
-import { comma, dash, identifier, slash } from "../lexer";
+import { comma, dash, identifier, slash } from "@/lexer";
 
-export class BaseParser extends CstParser {
+export class CronParser extends CstParser {
   constructor(vocabulary: TokenVocabulary, invokedByChild: boolean = false) {
     super(vocabulary);
     if (!invokedByChild) {
