+++ src/parser/index.ts
@@ -1,3 +1,3 @@
-export * from "./BaseParser";
+export * from "./CronParser";
 export * from "./JenkinsParser";
 export * from "./QuartzParser";
