+++ src/semantic/index.ts
@@ -1,5 +1,5 @@
 export { default as AbstractVisitor } from "./AbstractVisitor";
-export { default as BaseVisitor } from "./BaseVisitor";
+export { default as CronVisitor } from "./CronVisitor";
 export { default as JenkinsVisitor } from "./JenkinsVisitor";
 export { default as QuartzVisitor } from "./QuartzVisitor";
