+++ src/syntax/index.ts
@@ -1,2 +1,2 @@
-export * from "./base";
+export * from "./cron";
 export * from "./quartz";
