+++ test/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./*"]
+      "@/*": ["../src/*"]
     }
   },
   "include": ["./**/*.ts"]
